feat: refactor UV0 handling to use half precision types for improved performance

+++ Assets/Runtime/Engine/Jobs/Meshing/GreedyMesher.StructsAndHelpers.cs
@@ -18,7 +18,7 @@ namespace Runtime.Engine.Jobs.Meshing
         [BurstCompile]
         private struct UVQuad
         {
-            public float4 Uv1, Uv2, Uv3, Uv4;
+            public half4 Uv1, Uv2, Uv3, Uv4;
         }
 
         [BurstCompile]
