+++ Assets/Runtime/Engine/Jobs/Meshing/MeshBuildScheduler.cs
@@ -64,7 +64,7 @@ namespace Runtime.Engine.Jobs.Meshing
                 // Int interpolation cause issues
                 [0] = new VertexAttributeDescriptor(VertexAttribute.Position),
                 [1] = new VertexAttributeDescriptor(VertexAttribute.Normal),
-                [2] = new VertexAttributeDescriptor(VertexAttribute.TexCoord0, VertexAttributeFormat.Float32, 4),
+                [2] = new VertexAttributeDescriptor(VertexAttribute.TexCoord0, VertexAttributeFormat.Float16, 4),
                 [3] = new VertexAttributeDescriptor(VertexAttribute.TexCoord1, VertexAttributeFormat.Float32, 4),
                 [4] = new VertexAttributeDescriptor(VertexAttribute.TexCoord2, VertexAttributeFormat.Float32, 4)
             };
